Initial import of draft-ietf-dots-use-cases-04.
The markdown in draft-ietf-dots-use-cases.md results in a document in
text form with superficial differences against the published -04
revision: some capitalization changes, a handful of reference formatting
changes, and a minor correction to a reference to a section in the
document, all resulting from using the kramdown-rfc2629 utility.

The document otherwise should be unchanged from the published -04 draft.